requirement.txt
asgiref==3.5.0
cycler==0.11.0
Django==3.2.25
django-simpleui==2024.4.1
django-tinymce==3.6.1
fonttools== 4.29.1
joblib==1.3.2
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.21.5
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.7.3
setuptools==68.0.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
typing-extensions==4.7.1
tzdata==2021.5